SUMMARY:  removed specializations for size_t
     CHECKLIST
     ---------

     1) Did you 'cvs add' all new files and directories [Y:N:N/A]?  n/a

     2) Did you add new source files to a CMakeLists.txt file [Y:N:N/A]?  no

     3)  Did you modify or add any source or build system files?  yes
     3a) Did you rebuild the libraries and executables?  yes
     3b) Did you rerun the test suite on the packages that you changed and
         all packages that depend on the modified packages?  yes
     3c) Did you change the required or optional dependencies for any
         package?  n/a


     DESCRIPTION OF CHANGES:

Removed Teuchos specializations for "size_t" in OrdinalTraits, SerializationTraits and ScalarTraits.
Added specializations for my machine's size_t ("unsigned long")
Removed these specialization's from PyTrilinos to keep from conflicting.

Tests build/pass for Tpetra/Teuchos/Belos on my machine. I can't get PyTrilinos to build, so I'll check harness results tomorrow.

(checklist version 2009-07-14)

+++ packages/teuchos/src/Teuchos_SerializationTraits.hpp
@@ -212,8 +212,8 @@ class SerializationTraits<Ordinal,long int>
 {};
 
 template<typename Ordinal>
-class SerializationTraits<Ordinal,size_t>
-  : public DirectSerializationTraits<Ordinal,size_t>
+class SerializationTraits<Ordinal,long unsigned int>
+  : public DirectSerializationTraits<Ordinal,long unsigned int>
 {};
 
 template<typename Ordinal>
